Fix perf regression Windows + CMake 3.18+
The code that was converting from compile_commands.json to
compile_commands.json.bin was not properly stripping the name
of the output file. The result was that clang was called once
per output file rather than once per unique combination of
clang flags. Thanks to Emre for diagnosing this issue.

Bug: 196847363
Test: Added test with Emre's original repro compile_commands.json
Change-Id: If4deb2f3402ff1f39f2a260c2972a0e0bf4a3333