200068492: Remove extra entries in baselines with --write-reference-baseline
Also fixes the flag to work correctly when no input baseline
is specified.

Finally, updates the @RequireApi enforcement to make error messages
a bit more accurate in terms of what they apply to (calling a
method versus referencing a field versus extending a class). This is
important since baseline matching is based purely on messages within
a file.

Test: Unit tests included
Bug: 200068492
Change-Id: I65a7aedf28afa0fdf9f673f9efd4c793f03ca70d