Always put the notice file at the root
..in the bazel kotlin rule.

Our current convention for most maven published libraries is that the
notice file is adjacent to the build file, but resources are in
src/main/resources.

e.g.

tools/base/common
├── BUILD
├── NOTICE
├── src
│   ├── main
│   │   ├── resources

A single resource_strip_prefix can't model this, but as the notice file
is special-cased already, it makes ense to special-case its location at
the root of the jar too.

Bug: 143864616
Test: To follow
Change-Id: Ied91b97fb2478766fc4c1ce72813bf6f81a6c8e3