Add support for function literals to LiveEdit
Introduces the notion of a proxied class to LiveEdit. When a proxied
class is instantiated within the interpreter, LiveEdit will create a
proxy object implementing the same interfaces as the class.

This proxy object forwards method invocations and field access to a
ProxyClassHandler, which uses the LiveEdit interpreter to execute
further method invocations. This allows function objects to be created
and passed around inside an interpreted context.

Additionally, to support statically cached INSTANCE fields for
non-capturing lambdas within Kotlin and Compose, proxied classes also
execute <clinit> and allow static fields to be get/set.

To support synthetically generated accessor methods, special casing is
added to the interpreter to use LiveEdit-held bytecode when resolving
such methods. This allows the interpreter to call newly generated
methods that were not present when the original application was built.

Bug: None
Test: Manual
Change-Id: Ibe7eaa84945e79ce78e0d9ebca925c856b061835