[Hybrid build] Reduce AGP compilation dependencies
The Android Gradle Plugin no longer needs to compile against lint.

Hopefully this will mean a shorter critical path when recompiling lint
or AGP.

The development gradle build still needs the artifacts to compile and
run tests against, which is modelled in //tools/base:agp_test_artifacts.
Eventually it'd be good to split that further so the compilation can
happen in parallel, but there's further cleanup to do first.

Bug: 143864616
Test: GmavenZipTest, manually ran gradlew :publishLocal compileTestJava
Change-Id: I592a818eec99211f57a510b061f2ae005d55bab5