203452203: Unnecessary NewApi error at android.util.SparseArray#set
This CL fixes the following 2 bugs and begins work on a third:
203452203: Unnecessary NewApi error at android.util.SparseArray#set
177353340: getSuppressed requires API 19, but addSuppressed does not

R8 will always perform some basic desugaring for APIs like Long.compare,
Objects.requireNonNull, etc. Lint had a hardcoded list of these, but the
list was incomplete, and missed a number of recently added methods.

This CL updates lint's hardcoded list to the current set.

It also adds basic support (via a private command line flag,
`--Xdesugared-methods`) for AGP to set the set of desugared methods for
lint to use. This is to kickstart the work on passing the full backport
list to lint from the build system; issue 140368601.

Fixes: 177353340
Fixes: 203452203
Bug: 140368601
Test: Unit test included
Change-Id: Ie94a620bc03c36f2f26eba41ebc271746f7167de