Update GRADLE_PLUGIN_..._VERSION constants for reality
The minimum supported version of the Android Gradle Plugin in Studio is
primarily determined by sync; at present, as of Bumblebee and current
Chipmunk, that is the 3.2.0 series.

Nothing needs to change other than the tests of lint's checks, which
need updating because the lint error from using an unsupported version
of the Android Gradle Plugin in a project is mutually exclusive with
warnings related to use of outdated AGP versions.

Bug: 204703589
Test: existing tests updated
Change-Id: I63006c8c75f75d35bccd31027586ba03c743202c